primer_deploy/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
decouple==0.0.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
